Refactor integration tests to use real modules
