+++ tests/integration_modules/match_guard.py
@@ -1,0 +1,6 @@
+def probe(value):
+    match value:
+        case iterable if not hasattr(iterable, "__next__"):
+            return f"no next for {type(iterable).__name__}"
+        case _:
+            return "has next"
