+++ tests/integration_modules/translation_module.py
@@ -1,0 +1,13 @@
+_ = lambda text: f"translated:{text}"
+
+
+def translate_message():
+    _("before try")
+    try:
+        raise RuntimeError("boom")
+    except RuntimeError:
+        return _("after except")
+
+
+def call_translate():
+    return translate_message()
